requirements.txt
python-docx== 0.8.10
pyttsx3== 2.90
# to install the modules, type in the terminal "pip3 install -r requirements.txt"